Support line coordinates in "extended" SubRip text
According to Wikipedia, a sequence of X/Y coordinates may be included at
the end of a timestamp row to specify placement of individual lines. VLC
ignores this extra info, however.

The source (http://forum.doom9.org/archive/index.php/t-86664.html) which
describes this feature uses only integers in its examples; to err on the
side of caution, it should be assumed that floats are supported too.