Hide unrecognised SSA tags in SubRip Text files
This builds upon one of the tests in `./samples/srt/SSA Extensions.srt`,
which asserts that foreign SSA tags should be hidden instead of rendered
as subtitle text (behaviour consistent with VLC's interpretation). These
tags are now highlighted as errors.

    1
    00:00:01,000 --> 00:00:02,000
    - Hide: {Y:tag stuff}
    - Hide: {\ letters_or_numbers or spaces?}
    - Hide: { :whitespace accepted as tag-name}
    - Partially-hide: {Y:nested {braces} aren't supported}
    - Show: {XY:multi-character tag-names}
    - Show: {normal text}

The above example is displayed by VLC 3.0.20 as the subtitle lines would
suggest, with the fourth line ending with " aren't supported}".